pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "grading-rubric"
version = "0.1.0"
description = "Grading Rubric Studio — assesses and improves teacher rubrics against three quality criteria."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Wiktor Lisowski" }]
license = { text = "Proprietary" }

# Pinned per § 3 of docs/design.md. No `validance-sdk` here — DR-ARC-07.
dependencies = [
    "pydantic>=2.6,<3",
    "anthropic>=0.34",
    "openai>=1.40",
    "krippendorff>=0.6",
    "pypdf>=4.2",
    "pdfplumber>=0.11",
    "python-docx>=1.1",
    "Pillow>=10.0",
    "PyYAML>=6.0",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
    "mypy>=1.10",
]

[project.scripts]
grading-rubric-cli = "grading_rubric.cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["grading_rubric*"]
exclude = ["tests*", "validance*", "frontend*"]

[tool.setuptools.package-data]
"grading_rubric.gateway" = ["prompts/*.md"]

[tool.ruff]
line-length = 110
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
